Move hook to seperate file
